Added project to repo
This is just to add all the training files I have made to a github repo. I might want to access them on different computers since this is like a sort of notes to remember some of the most used classes in bootstrap and how to add custom styles to bootstrap classes.